code to create a blocks added
Removed some screen clearing code and added the lines to display a new
block when the user clicks.  It doesn't yet keep track of the block,
we'll need an aray for that.  We also need a screen clearing function so
we can destroy blocks as well as create them.  This is different from
the block destroying function which we also need to create.